Create task refreshVersionsCleanup
